+++ sync-for-payables/src/codat_sync_for_payables/sdk.py
@@ -28,6 +28,16 @@ class CodatSyncPayables(BaseSDK):
     [Explore product](https://docs.codat.io/payables/bill-pay-kit) | [See OpenAPI spec](https://github.com/codatio/oas)
 
     ---
+    ## Supported Integrations
+
+    | Integration                   | Supported |
+    |-------------------------------|-----------|
+    | FreeAgent                     | Yes       |
+    | QuickBooks Online             | Yes       |
+    | Oracle NetSuite               | Yes       |
+    | Xero                          | Yes       |
+
+    ---
     <!-- Start Codat Tags Table -->
     ## Endpoints