+++ sync-for-payables/RELEASES.md
@@ -9,3 +9,13 @@ Based on:
 - [python v4.0.0] sync-for-payables
 ### Releases
 - [PyPI v4.0.0] https://pypi.org/project/codat-sync-for-payables/4.0.0 - sync-for-payables
+
+## 2024-10-29 14:20:03
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.423.1 (2.442.11) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v5.0.0] sync-for-payables
+### Releases
+- [PyPI v5.0.0] https://pypi.org/project/codat-sync-for-payables/5.0.0 - sync-for-payables